Move message-support to email controllers
