+++ server/routes.js
@@ -26,7 +26,6 @@ import graphqlSchemaV2 from './graphql/v2/schema';
 import cache from './lib/cache';
 import errors from './lib/errors';
 import logger from './lib/logger';
-import sendMessageToSupport from './lib/message-support';
 import oauth, { authorizeAuthenticateHandler } from './lib/oauth';
 import { reportMessageToSentry, SentryGraphQLPlugin } from './lib/sentry';
 import { parseToBoolean } from './lib/utils';
@@ -267,7 +266,7 @@ export default async app => {
   /**
    * Contact Form
    */
-  app.post('/contact/send-message', sendMessageToSupport);
+  app.post('/contact/send-message', email.messageSupport);
 
   app.use(sanitizer()); // note: this break /webhooks/mailgun /graphiql
